[APP4-HOOKS] useState hook
Note that if you use functional update (https://reactjs.org/docs/hooks-reference.html#functional-updates) in the onChange closure,
the functional update is an inner closure.
The event from the outer closure is locked in the inner closure.
Unfortunately, due to the fact that react uses SyntheticEvents and event pooling  the SyntheticEvent object will be reused and all properties will be nullified after the onChange event callback has been invoked.
 As such, you cannot access the event in an asynchronous way. (like in the inner setState) See https://reactjs.org/docs/events.html#event-pooling